HSEARCH-3438 Make BuiltinContainerExtractor instances internal only
