+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/extractor/builtin/impl/IterableElementExtractor.java
@@ -4,16 +4,16 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.extractor.builtin;
+package org.hibernate.search.mapper.pojo.extractor.builtin.impl;
 
-import java.util.Arrays;
 import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
 
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractor;
 
-public class ArrayElementExtractor<T> implements ContainerExtractor<T[], T> {
+public class IterableElementExtractor<T> implements ContainerExtractor<Iterable<T>, T> {
 	@Override
-	public Stream<T> extract(T[] container) {
-		return container == null ? Stream.empty() : Arrays.stream( container );
+	public Stream<T> extract(Iterable<T> container) {
+		return container == null ? Stream.empty() : StreamSupport.stream( container.spliterator(), false );
 	}
 }
