+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/extractor/builtin/impl/MapKeyExtractor.java
@@ -4,16 +4,16 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.extractor.builtin;
+package org.hibernate.search.mapper.pojo.extractor.builtin.impl;
 
 import java.util.Map;
 import java.util.stream.Stream;
 
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractor;
 
-public class MapValueExtractor<T> implements ContainerExtractor<Map<?, T>, T> {
+public class MapKeyExtractor<T> implements ContainerExtractor<Map<T, ?>, T> {
 	@Override
-	public Stream<T> extract(Map<?, T> container) {
-		return container == null ? Stream.empty() : container.values().stream();
+	public Stream<T> extract(Map<T, ?> container) {
+		return container == null ? Stream.empty() : container.keySet().stream();
 	}
 }
