+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/extractor/builtin/impl/OptionalDoubleValueExtractor.java
@@ -4,18 +4,18 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.pojo.extractor.builtin;
+package org.hibernate.search.mapper.pojo.extractor.builtin.impl;
 
-import java.util.OptionalInt;
+import java.util.OptionalDouble;
 import java.util.stream.Stream;
 
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractor;
 
-public class OptionalIntValueExtractor implements ContainerExtractor<OptionalInt, Integer> {
+public class OptionalDoubleValueExtractor implements ContainerExtractor<OptionalDouble, Double> {
 	@Override
-	public Stream<Integer> extract(OptionalInt container) {
+	public Stream<Double> extract(OptionalDouble container) {
 		if ( container != null && container.isPresent() ) {
-			return Stream.of( container.getAsInt() );
+			return Stream.of( container.getAsDouble() );
 		}
 		else {
 			return Stream.empty();
