+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/extractor/impl/ContainerExtractorBinder.java
@@ -19,14 +19,14 @@ import org.hibernate.search.engine.environment.bean.BeanProvider;
 import org.hibernate.search.engine.mapper.mapping.spi.MappingBuildContext;
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractor;
 import org.hibernate.search.mapper.pojo.extractor.ContainerExtractorPath;
-import org.hibernate.search.mapper.pojo.extractor.builtin.ArrayElementExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.CollectionElementExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.IterableElementExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.MapValueExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.OptionalDoubleValueExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.OptionalIntValueExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.OptionalLongValueExtractor;
-import org.hibernate.search.mapper.pojo.extractor.builtin.OptionalValueExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.ArrayElementExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.CollectionElementExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.IterableElementExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.MapValueExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.OptionalDoubleValueExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.OptionalIntValueExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.OptionalLongValueExtractor;
+import org.hibernate.search.mapper.pojo.extractor.builtin.impl.OptionalValueExtractor;
 import org.hibernate.search.mapper.pojo.logging.impl.Log;
 import org.hibernate.search.mapper.pojo.model.spi.PojoGenericTypeModel;
 import org.hibernate.search.mapper.pojo.model.typepattern.impl.ExtractingTypePatternMatcher;
